+++ packages/i18nify-react/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @razorpay/i18nify-react
 
+## 1.1.0
+
+### Minor Changes
+
+- 7ced55a: [feat]: update @razorpay/i18nify to @razorpay/i18nify-js
+
+## 1.0.17
+
+### Patch Changes
+
+- e538b9b: [feat]: initData support added for I18nProvider
+
+## 1.0.16
+
+### Patch Changes
+
+- 50aca2e: [fix]: package.json versioning fixed
+
 ## 1.0.3
 
 ### Patch Changes